--- src/Map.tsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import logo from './logo.svg';
-import './App.css';
-
-
-
-const Map = () => {
-  return (
-   
-    <h1>Mapa</h1>
-  );
-}
-
-
-
-export default Map;
