+++ src/pages/map/Map.css
@@ -1,0 +1,5 @@
+.map-container  {
+    width: 1200px;
+    height: 1200px;
+    background-color: red;
+}
